Split out trailers and reset tests #28306
